Modularized services and components for better code readability

+++ src/main/java/com/score_me/was_metrics_exporter/controllers/PgMetricsController.java
@@ -24,7 +24,7 @@ public class PgMetricsController {
             return ResponseEntity.badRequest().body("Missing processGroupId");
         }
 
-        Map<String, Long> metrics = pgMetricsService.getMetricsForGroup(groupId);
+        Map<String, Double> metrics = pgMetricsService.getMetricsForGroup(groupId);
         return ResponseEntity.ok(metrics);
     }
 }
